refactor(dh): move YesNo pipe to shared folder
